wanderapp_backend/requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
Django==4.2.23
django-cors-headers==4.7.0
django-countries==7.6.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
gpxpy==1.6.2
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing-extensions==4.15.0
